# perf: some more python overhead reduction