This PR reduces some of the python overhead (it's a continued investigation of https://github.com/scikit-hep/awkward/pull/3359).

With this PR the runtimes can be reduced further by ~10% (based on the same example of https://github.com/scikit-hep/awkward/pull/3359) to:
- CPU backend: 25.1 ms ± 126 μs -> **23.4 ms ± 269 μs**
- TypeTracer backend: 24.4 ms ± 195 μs -> **22.3 ms ± 130 μs**

The main improvements are:
- Merging 2 for loops in `broadcast_any_list()`
- Merging 2 for loops in `TypeTracer.broadcast_arrays()`
- Avoid unnecessary `n_missing_dims` recalculations in `TypeTracer.__getitem__` loop
- Avoid named_axis propagation in slicing when no named_axis are present
- Add `nplikes` to `Index` instance calls where possible to avoid runtime type checks and nplike inference with (`nplike_of_obj`)